refactor(scripts): update latency script
